Refactor web UI assets into modular components
