+++ baseball_sim/ui/web/static/css/base.css
@@ -1,0 +1,347 @@
+:root {
+  --bg: #020617;
+  --surface: rgba(12, 20, 40, 0.92);
+  --surface-strong: rgba(4, 10, 24, 0.96);
+  --surface-soft: rgba(15, 23, 42, 0.75);
+  --accent: #22d3ee;
+  --accent-muted: #38bdf8;
+  --highlight: #a855f7;
+  --text: #f8fafc;
+  --text-muted: rgba(191, 219, 254, 0.75);
+  --danger: #f87171;
+  --success: #34d399;
+  --warning: #facc15;
+  --border: rgba(56, 189, 248, 0.18);
+  --grid-line: rgba(56, 189, 248, 0.12);
+  --glow: rgba(56, 189, 248, 0.32);
+  --glow-strong: rgba(168, 85, 247, 0.32);
+  --font-body: 'Noto Sans JP', 'Segoe UI', 'Roboto', sans-serif;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  font-family: var(--font-body);
+  background:
+    radial-gradient(circle at 18% 24%, rgba(56, 189, 248, 0.2), transparent 45%),
+    radial-gradient(circle at 82% 12%, rgba(168, 85, 247, 0.18), transparent 58%),
+    radial-gradient(circle at 50% 120%, rgba(15, 118, 110, 0.15), transparent 65%),
+    #020617;
+  background-attachment: fixed;
+  color: var(--text);
+  min-height: 100vh;
+  position: relative;
+  overflow-x: hidden;
+}
+
+body::before {
+  content: '';
+  position: fixed;
+  inset: 0;
+  background:
+    linear-gradient(90deg, transparent 0%, var(--grid-line) 49%, transparent 51%),
+    linear-gradient(0deg, transparent 0%, var(--grid-line) 49%, transparent 51%);
+  background-size: 120px 120px;
+  opacity: 0.35;
+  pointer-events: none;
+  z-index: 0;
+}
+
+body::after {
+  content: '';
+  position: fixed;
+  inset: -20% -30%;
+  background: radial-gradient(circle at center, rgba(56, 189, 248, 0.12), transparent 55%);
+  filter: blur(60px);
+  pointer-events: none;
+  z-index: 0;
+}
+
+.hidden {
+  display: none !important;
+}
+
+.visually-hidden {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  border: 0;
+}
+
+.app-shell {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  z-index: 1;
+}
+
+.app-header {
+  position: relative;
+  padding: 32px clamp(20px, 6vw, 48px) 28px;
+  border-bottom: 1px solid var(--border);
+  overflow: hidden;
+  background: linear-gradient(145deg, rgba(34, 211, 238, 0.08), rgba(15, 23, 42, 0.85));
+}
+
+.header-overlay {
+  position: absolute;
+  inset: -40% -20%;
+  background:
+    radial-gradient(circle at 20% 20%, rgba(56, 189, 248, 0.25), transparent 55%),
+    radial-gradient(circle at 80% 10%, rgba(168, 85, 247, 0.22), transparent 60%),
+    radial-gradient(circle at 60% 80%, rgba(14, 165, 233, 0.18), transparent 65%);
+  filter: blur(40px);
+  opacity: 0.9;
+  pointer-events: none;
+  z-index: 0;
+}
+
+.header-content {
+  position: relative;
+  z-index: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  max-width: 1200px;
+  margin: 0 auto;
+  width: 100%;
+}
+
+.header-topline {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: flex-end;
+  justify-content: space-between;
+  gap: 24px;
+}
+
+.title-group {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  align-items: flex-start;
+}
+
+.title-group h1 {
+  margin: 0;
+  font-size: clamp(32px, 5vw, 48px);
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+  text-shadow: 0 6px 16px rgba(8, 47, 73, 0.45);
+}
+
+.subtitle {
+  margin: 12px 0 0;
+  color: var(--text-muted);
+  max-width: 640px;
+  font-size: 15px;
+  letter-spacing: 0.04em;
+}
+
+.header-tags {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 12px;
+  align-items: center;
+  justify-content: flex-end;
+}
+
+.header-tag {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  padding: 6px 14px;
+  border-radius: 999px;
+  background: rgba(56, 189, 248, 0.12);
+  border: 1px solid rgba(56, 189, 248, 0.35);
+  color: var(--accent-muted);
+  font-size: 11px;
+  letter-spacing: 0.18em;
+  text-transform: uppercase;
+  box-shadow:
+    0 10px 24px rgba(8, 47, 73, 0.35),
+    inset 0 0 12px rgba(56, 189, 248, 0.15);
+}
+
+.header-insights {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+  gap: 16px;
+}
+
+.insight-card {
+  position: relative;
+  padding: 18px 20px;
+  border-radius: 18px;
+  border: 1px solid rgba(56, 189, 248, 0.25);
+  background: linear-gradient(140deg, rgba(8, 11, 24, 0.75), rgba(8, 47, 73, 0.18));
+  box-shadow: 0 18px 40px rgba(2, 6, 23, 0.35);
+  overflow: hidden;
+  backdrop-filter: blur(8px);
+}
+
+.insight-card::before {
+  content: '';
+  position: absolute;
+  inset: 0;
+  background: radial-gradient(circle at top right, rgba(168, 85, 247, 0.28), transparent 55%);
+  opacity: 0.8;
+  pointer-events: none;
+}
+
+.insight-card::after {
+  content: '';
+  position: absolute;
+  top: -120%;
+  left: -60%;
+  width: 220%;
+  height: 220%;
+  background: linear-gradient(130deg, transparent 40%, rgba(56, 189, 248, 0.15) 50%, transparent 60%);
+  transform: rotate(8deg);
+  animation: insight-glimmer 8s ease-in-out infinite;
+  pointer-events: none;
+}
+
+@keyframes insight-glimmer {
+  0% {
+    transform: translateX(-40%) rotate(8deg);
+    opacity: 0.1;
+  }
+  50% {
+    transform: translateX(30%) rotate(8deg);
+    opacity: 0.4;
+  }
+  100% {
+    transform: translateX(100%) rotate(8deg);
+    opacity: 0.05;
+  }
+}
+
+.insight-header {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+  margin-bottom: 12px;
+  font-size: 12px;
+  letter-spacing: 0.18em;
+  text-transform: uppercase;
+  color: var(--text-muted);
+}
+
+.insight-icon {
+  font-size: 18px;
+  opacity: 0.8;
+}
+
+.insight-value {
+  position: relative;
+  font-size: clamp(28px, 4vw, 40px);
+  font-weight: 700;
+  margin: 0;
+  color: var(--accent-muted);
+  text-shadow: 0 10px 24px rgba(8, 47, 73, 0.65);
+}
+
+.insight-value[data-trend='positive'] {
+  color: var(--success);
+}
+
+.insight-value[data-trend='negative'] {
+  color: var(--danger);
+}
+
+.insight-card[data-insight='base-pressure'] .insight-value[data-intensity='high'] {
+  color: var(--warning);
+}
+
+.insight-card[data-insight='base-pressure'] .insight-value[data-intensity='medium'] {
+  color: var(--accent-muted);
+}
+
+.insight-card[data-insight='base-pressure'] .insight-value[data-intensity='low'] {
+  color: var(--text-muted);
+}
+
+.insight-caption {
+  margin: 6px 0 0;
+  font-size: 13px;
+  color: var(--text-muted);
+  letter-spacing: 0.04em;
+}
+
+.insight-meta {
+  margin: 10px 0 0;
+  font-size: 12px;
+  letter-spacing: 0.08em;
+  color: rgba(148, 163, 184, 0.75);
+}
+
+.insight-meter {
+  margin-top: 12px;
+  display: grid;
+  gap: 6px;
+}
+
+.insight-meter-track {
+  position: relative;
+  width: 100%;
+  height: 6px;
+  border-radius: 999px;
+  background: rgba(148, 163, 184, 0.18);
+  overflow: hidden;
+}
+
+.insight-meter-fill {
+  position: absolute;
+  inset: 0;
+  width: 0%;
+  border-radius: inherit;
+  background: linear-gradient(90deg, var(--accent) 0%, var(--highlight) 100%);
+  box-shadow:
+    0 0 12px var(--glow),
+    0 0 24px var(--glow-strong);
+  transition: width 0.6s ease;
+}
+
+.insight-meter-label {
+  font-size: 12px;
+  color: rgba(191, 219, 254, 0.85);
+  letter-spacing: 0.1em;
+}
+
+.developer-hint {
+  margin-top: 8px;
+  opacity: 0.75;
+  transition: opacity 0.3s ease;
+  font-size: 13px;
+}
+
+.developer-hint:hover {
+  opacity: 1;
+}
+
+.developer-hint small {
+  color: var(--text-muted);
+}
+
+.developer-hint kbd {
+  background: rgba(15, 23, 42, 0.65);
+  border: 1px solid rgba(56, 189, 248, 0.45);
+  border-radius: 6px;
+  padding: 2px 6px;
+  font-family: monospace;
+  font-size: 0.9em;
+  color: var(--accent-muted);
+  box-shadow: 0 1px 10px rgba(2, 6, 23, 0.45);
+}
