+++ baseball_sim/ui/web/static/css/layout.css
@@ -1,0 +1,272 @@
+.app-main {
+  flex: 1;
+  padding: 32px clamp(20px, 6vw, 56px) 72px;
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+  position: relative;
+  max-width: 1320px;
+  margin: 0 auto;
+  width: 100%;
+}
+
+.app-main::before {
+  content: '';
+  position: absolute;
+  inset: 40px 10% -60px;
+  background: radial-gradient(circle at top, rgba(56, 189, 248, 0.12), transparent 65%);
+  filter: blur(80px);
+  z-index: -1;
+  pointer-events: none;
+}
+
+.status-message {
+  min-height: 32px;
+  color: var(--accent-muted);
+  font-weight: 600;
+  padding: 12px 18px;
+  border-radius: 12px;
+  background: rgba(8, 11, 24, 0.6);
+  border: 1px solid rgba(56, 189, 248, 0.25);
+  box-shadow: 0 18px 36px rgba(2, 6, 23, 0.45);
+  transition: all 0.3s ease;
+}
+
+.status-message.danger {
+  color: var(--danger);
+}
+
+.status-message.success {
+  color: var(--success);
+  background: rgba(52, 211, 153, 0.12);
+  border-color: rgba(52, 211, 153, 0.35);
+  box-shadow: 0 18px 42px rgba(16, 185, 129, 0.25);
+}
+
+.status-message.info {
+  color: var(--accent);
+}
+
+.status-message.warning {
+  color: var(--warning);
+  background: rgba(250, 204, 21, 0.12);
+  border-color: rgba(250, 204, 21, 0.32);
+  box-shadow: 0 16px 32px rgba(250, 204, 21, 0.18);
+}
+
+.screen.hidden {
+  display: none;
+}
+
+.title-card {
+  position: relative;
+  background: linear-gradient(150deg, rgba(4, 10, 24, 0.92), rgba(8, 47, 73, 0.45));
+  border: 1px solid rgba(56, 189, 248, 0.25);
+  border-radius: 24px;
+  padding: 32px clamp(20px, 5vw, 48px);
+  box-shadow:
+    0 36px 80px rgba(2, 6, 23, 0.6),
+    inset 0 0 0 1px rgba(148, 163, 184, 0.08);
+  max-width: 980px;
+  margin: 0 auto;
+  overflow: hidden;
+}
+
+.title-card::before {
+  content: '';
+  position: absolute;
+  inset: 0;
+  background:
+    linear-gradient(125deg, rgba(56, 189, 248, 0.12), transparent 60%),
+    repeating-linear-gradient(
+      0deg,
+      transparent,
+      transparent 14px,
+      rgba(56, 189, 248, 0.08) 14px,
+      rgba(56, 189, 248, 0.08) 15px
+    );
+  opacity: 0.4;
+  pointer-events: none;
+}
+
+.title-card > * {
+  position: relative;
+  z-index: 1;
+}
+
+.title-card::after {
+  content: '';
+  position: absolute;
+  inset: 16px;
+  border-radius: 20px;
+  border: 1px solid rgba(148, 163, 184, 0.12);
+  opacity: 0.5;
+  pointer-events: none;
+}
+
+.title-card h2 {
+  margin-top: 0;
+  letter-spacing: 0.12em;
+  text-transform: uppercase;
+  font-size: 20px;
+}
+
+.team-status-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
+  gap: 18px;
+  margin: 20px 0 28px;
+}
+
+.team-card {
+  position: relative;
+  background: rgba(8, 11, 24, 0.78);
+  border: 1px solid rgba(56, 189, 248, 0.2);
+  border-radius: 16px;
+  padding: 18px;
+  min-height: 170px;
+  box-shadow: 0 22px 48px rgba(2, 6, 23, 0.5);
+  overflow: hidden;
+}
+
+.team-card > * {
+  position: relative;
+  z-index: 1;
+}
+
+.team-card::before {
+  content: '';
+  position: absolute;
+  inset: 0;
+  background: linear-gradient(135deg, rgba(56, 189, 248, 0.08), transparent 65%);
+  opacity: 0.6;
+  pointer-events: none;
+}
+
+.team-card h3 {
+  margin: 0 0 8px;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+}
+
+.team-message {
+  font-weight: 600;
+  color: var(--accent-muted);
+}
+
+.team-errors {
+  margin: 12px 0 0;
+  padding-left: 20px;
+  color: var(--warning);
+  font-size: 14px;
+}
+
+.title-hint {
+  margin: 0 0 16px;
+  color: var(--text-muted);
+  letter-spacing: 0.03em;
+}
+
+.title-actions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 12px;
+}
+
+button {
+  border: 1px solid rgba(56, 189, 248, 0.32);
+  border-radius: 999px;
+  padding: 12px 22px;
+  background: rgba(8, 11, 24, 0.65);
+  color: var(--text);
+  font-size: 15px;
+  font-weight: 600;
+  letter-spacing: 0.08em;
+  cursor: pointer;
+  transition: transform 0.2s ease, background 0.2s ease, box-shadow 0.2s ease, border-color 0.2s ease;
+  box-shadow: 0 14px 30px rgba(2, 6, 23, 0.45);
+  backdrop-filter: blur(4px);
+}
+
+button.primary {
+  background: linear-gradient(135deg, var(--accent) 0%, var(--highlight) 100%);
+  color: #020617;
+  border: none;
+  box-shadow:
+    0 16px 36px rgba(34, 211, 238, 0.35),
+    0 0 18px rgba(168, 85, 247, 0.25);
+}
+
+button:hover:not(:disabled) {
+  transform: translateY(-2px);
+  background: rgba(15, 23, 42, 0.75);
+  border-color: rgba(56, 189, 248, 0.5);
+  box-shadow:
+    0 18px 40px rgba(2, 6, 23, 0.55),
+    0 0 12px rgba(56, 189, 248, 0.2);
+}
+
+button.primary:hover:not(:disabled) {
+  background: linear-gradient(135deg, var(--accent) 0%, rgba(168, 85, 247, 0.92) 100%);
+  box-shadow:
+    0 22px 44px rgba(34, 211, 238, 0.45),
+    0 0 20px rgba(168, 85, 247, 0.4);
+}
+
+button:disabled {
+  opacity: 0.45;
+  cursor: not-allowed;
+  box-shadow: none;
+}
+
+button:focus-visible {
+  outline: none;
+  box-shadow: 0 0 0 3px rgba(34, 211, 238, 0.35);
+}
+
+.toolbar {
+  display: flex;
+  justify-content: space-between;
+  gap: 16px;
+  padding: 20px 24px;
+  background: var(--surface-soft);
+  border: 1px solid rgba(56, 189, 248, 0.2);
+  border-radius: 20px;
+  margin-bottom: 20px;
+  box-shadow: 0 24px 56px rgba(2, 6, 23, 0.5);
+  backdrop-filter: blur(6px);
+}
+
+.toolbar-left,
+.toolbar-right {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+.action-warning {
+  color: var(--warning);
+  font-weight: 600;
+}
+
+.log-hint {
+  color: var(--text-muted);
+  font-size: 0.85em;
+  margin-right: 16px;
+  opacity: 0.6;
+  transition: opacity 0.3s ease;
+}
+
+.log-hint:hover {
+  opacity: 1;
+}
+
+.log-hint kbd {
+  background: rgba(15, 23, 42, 0.6);
+  border: 1px solid rgba(56, 189, 248, 0.4);
+  border-radius: 4px;
+  padding: 2px 6px;
+  font-family: monospace;
+  font-size: 0.9em;
+  color: var(--accent-muted);
+}
