+++ baseball_sim/ui/web/static/js/controllers/events.js
@@ -1,0 +1,148 @@
+import { elements } from '../dom.js';
+import { stateCache } from '../state.js';
+import {
+  hideDefenseMenu,
+  hideOffenseMenu,
+  toggleDefenseMenu,
+  toggleLogPanel,
+  toggleOffenseMenu,
+} from '../ui/menus.js';
+import { closeModal, openModal, resolveModal } from '../ui/modals.js';
+import { updateStatsPanel } from '../ui/renderers.js';
+import { handleDefenseBenchClick, handleDefenseFieldClick, updateDefenseSelectionInfo } from '../ui/defensePanel.js';
+
+export function initEventListeners(actions) {
+  elements.startButton.addEventListener('click', () => actions.handleStart(false));
+  elements.reloadTeams.addEventListener('click', actions.handleReloadTeams);
+  elements.restartButton.addEventListener('click', () => actions.handleStart(true));
+  elements.returnTitle.addEventListener('click', actions.handleReturnToTitle);
+  elements.clearLog.addEventListener('click', actions.handleClearLog);
+  elements.swingButton.addEventListener('click', actions.handleSwing);
+  elements.buntButton.addEventListener('click', actions.handleBunt);
+
+  if (elements.pinchButton) {
+    elements.pinchButton.addEventListener('click', actions.handlePinchHit);
+  }
+  if (elements.openOffenseButton) {
+    elements.openOffenseButton.addEventListener('click', toggleOffenseMenu);
+  }
+  if (elements.offensePinchMenuButton) {
+    elements.offensePinchMenuButton.addEventListener('click', () => openModal('offense'));
+  }
+  if (elements.openDefenseButton) {
+    elements.openDefenseButton.addEventListener('click', toggleDefenseMenu);
+  }
+  if (elements.defenseSubMenuButton) {
+    elements.defenseSubMenuButton.addEventListener('click', () => {
+      updateDefenseSelectionInfo();
+      openModal('defense');
+    });
+  }
+  if (elements.pitcherMenuButton) {
+    elements.pitcherMenuButton.addEventListener('click', () => openModal('pitcher'));
+  }
+  if (elements.openStatsButton) {
+    elements.openStatsButton.addEventListener('click', () => {
+      updateStatsPanel(stateCache.data);
+      openModal('stats');
+    });
+  }
+  if (elements.defenseApplyButton) {
+    elements.defenseApplyButton.addEventListener('click', actions.handleDefenseSubstitution);
+  }
+  if (elements.pitcherButton) {
+    elements.pitcherButton.addEventListener('click', actions.handlePitcherChange);
+  }
+
+  elements.modalCloseButtons.forEach((button) => {
+    const target = button.dataset.close;
+    button.addEventListener('click', () => closeModal(target || button.closest('.modal')));
+  });
+
+  ['offense', 'defense', 'pitcher', 'stats'].forEach((name) => {
+    const modal = resolveModal(name);
+    if (modal) {
+      modal.addEventListener('click', (event) => {
+        if (event.target === modal) {
+          closeModal(modal);
+        }
+      });
+    }
+  });
+
+  if (elements.offenseMenu) {
+    document.addEventListener('click', (event) => {
+      if (elements.offenseMenu.classList.contains('hidden')) {
+        return;
+      }
+      const clickedInsideMenu = elements.offenseMenu.contains(event.target);
+      const clickedToggle =
+        elements.openOffenseButton && elements.openOffenseButton.contains(event.target);
+      if (!clickedInsideMenu && !clickedToggle) {
+        hideOffenseMenu();
+      }
+    });
+  }
+
+  if (elements.defenseMenu) {
+    document.addEventListener('click', (event) => {
+      if (elements.defenseMenu.classList.contains('hidden')) {
+        return;
+      }
+      const clickedInsideMenu = elements.defenseMenu.contains(event.target);
+      const clickedToggle =
+        elements.openDefenseButton && elements.openDefenseButton.contains(event.target);
+      if (!clickedInsideMenu && !clickedToggle) {
+        hideDefenseMenu();
+      }
+    });
+  }
+
+  if (elements.defenseField) {
+    elements.defenseField.addEventListener('click', handleDefenseFieldClick);
+  }
+  if (elements.defenseBench) {
+    elements.defenseBench.addEventListener('click', handleDefenseBenchClick);
+  }
+  if (elements.defenseExtras) {
+    elements.defenseExtras.addEventListener('click', handleDefenseBenchClick);
+  }
+
+  elements.statsTeamButtons.forEach((button) => {
+    button.addEventListener('click', () => {
+      if (button.disabled) return;
+      const teamKey = button.dataset.statsTeam;
+      if (!teamKey) return;
+      stateCache.statsView.team = teamKey;
+      updateStatsPanel(stateCache.data);
+    });
+  });
+
+  elements.statsTypeButtons.forEach((button) => {
+    button.addEventListener('click', () => {
+      const viewType = button.dataset.statsType;
+      if (!viewType) return;
+      stateCache.statsView.type = viewType;
+      updateStatsPanel(stateCache.data);
+    });
+  });
+
+  document.addEventListener('keydown', (event) => {
+    if (event.key === 'Escape') {
+      hideOffenseMenu();
+      hideDefenseMenu();
+      ['offense', 'defense', 'pitcher', 'stats'].forEach((name) => {
+        const modal = resolveModal(name);
+        if (modal && !modal.classList.contains('hidden')) {
+          closeModal(modal);
+        }
+      });
+    }
+    if (event.key === 'Tab' && !event.ctrlKey && !event.altKey && !event.shiftKey) {
+      event.preventDefault();
+      toggleLogPanel();
+    }
+  });
+
+  updateDefenseSelectionInfo();
+}
