+++ baseball_sim/ui/web/static/js/dom.js
@@ -1,0 +1,68 @@
+// Centralised DOM element references used throughout the UI.
+
+export const elements = {
+  titleScreen: document.getElementById('title-screen'),
+  gameScreen: document.getElementById('game-screen'),
+  statusMessage: document.getElementById('status-message'),
+  startButton: document.getElementById('start-game'),
+  reloadTeams: document.getElementById('reload-teams'),
+  restartButton: document.getElementById('restart-game'),
+  returnTitle: document.getElementById('return-title'),
+  clearLog: document.getElementById('clear-log'),
+  swingButton: document.getElementById('swing-button'),
+  buntButton: document.getElementById('bunt-button'),
+  openOffenseButton: document.getElementById('open-offense-strategy'),
+  openDefenseButton: document.getElementById('open-defense-strategy'),
+  openStatsButton: document.getElementById('open-stats'),
+  offenseMenu: document.getElementById('offense-strategy-menu'),
+  offensePinchMenuButton: document.getElementById('open-offense-pinch-menu'),
+  defenseMenu: document.getElementById('defense-strategy-menu'),
+  defenseSubMenuButton: document.getElementById('open-defense-sub'),
+  pitcherMenuButton: document.getElementById('open-pitcher-change-menu'),
+  actionWarning: document.getElementById('action-warning'),
+  titleHint: document.getElementById('title-hint'),
+  logPanel: document.getElementById('log-panel'),
+  logContainer: document.getElementById('log-entries'),
+  scoreboard: document.getElementById('scoreboard'),
+  situationText: document.getElementById('situation-text'),
+  matchupText: document.getElementById('matchup-text'),
+  halfIndicator: document.getElementById('half-indicator'),
+  outsIndicator: document.getElementById('outs-indicator'),
+  defenseErrors: document.getElementById('defense-errors'),
+  offenseRoster: document.querySelector('#offense-roster tbody'),
+  defenseRoster: document.querySelector('#defense-roster tbody'),
+  offenseBench: document.getElementById('offense-bench'),
+  defenseBenchList: document.getElementById('defense-bench'),
+  homePitchers: document.querySelector('#home-pitchers ul'),
+  awayPitchers: document.querySelector('#away-pitchers ul'),
+  baseState: document.getElementById('base-state'),
+  pinchPlayer: document.getElementById('pinch-player'),
+  pinchButton: document.getElementById('pinch-hit-button'),
+  pinchCurrentBatter: document.getElementById('pinch-current-batter'),
+  offenseModal: document.getElementById('offense-modal'),
+  defenseModal: document.getElementById('defense-modal'),
+  pitcherModal: document.getElementById('pitcher-modal'),
+  statsModal: document.getElementById('stats-modal'),
+  modalCloseButtons: Array.from(document.querySelectorAll('.modal-close')),
+  defenseApplyButton: document.getElementById('defense-sub-button'),
+  defenseField: document.getElementById('defense-field'),
+  defenseBench: document.getElementById('defense-bench-panel'),
+  defenseExtras: document.getElementById('defense-extras'),
+  defenseSelectionInfo: document.getElementById('defense-selection-info'),
+  pitcherSelect: document.getElementById('pitcher-select'),
+  pitcherButton: document.getElementById('change-pitcher-button'),
+  statsTeamButtons: Array.from(document.querySelectorAll('[data-stats-team]')),
+  statsTypeButtons: Array.from(document.querySelectorAll('[data-stats-type]')),
+  statsTableHead: document.querySelector('#stats-table thead tr'),
+  statsTableBody: document.querySelector('#stats-table tbody'),
+  statsTitle: document.getElementById('stats-title'),
+  insightRunRate: document.getElementById('insight-run-rate'),
+  insightInningsSample: document.getElementById('insight-innings-sample'),
+  insightBasePressure: document.getElementById('insight-base-pressure'),
+  insightBaseCount: document.getElementById('insight-base-count'),
+  insightRunDiff: document.getElementById('insight-run-diff'),
+  insightProgressFill: document.getElementById('insight-progress-fill'),
+  insightProgressLabel: document.getElementById('insight-progress-label'),
+  insightMeter: document.querySelector('.insight-meter'),
+  insightGrid: document.getElementById('insight-grid'),
+};
